switch mainList to whatever user selects in the sidebar(friends, request, and followers))
