+++ src/js/react/FriendsPage.js
@@ -3,46 +3,31 @@ import { styled, createTheme, ThemeProvider } from '@mui/material/styles';
 import CssBaseline from '@mui/material/CssBaseline';
 import MuiDrawer from '@mui/material/Drawer';
 import Box from '@mui/material/Box';
-import MuiAppBar from '@mui/material/AppBar';
 import Toolbar from '@mui/material/Toolbar';
 import List from '@mui/material/List';
 import Typography from '@mui/material/Typography';
 import Divider from '@mui/material/Divider';
 import IconButton from '@mui/material/IconButton';
-import Badge from '@mui/material/Badge';
 import Container from '@mui/material/Container';
 import Grid from '@mui/material/Grid';
 import Paper from '@mui/material/Paper';
 import Link from '@mui/material/Link';
-import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
-import NotificationsIcon from '@mui/icons-material/Notifications';
-// import { mainListItems } from './FriendsUI/listItems';
-import Deposits from './FriendsUI/Deposits';
-import Orders from './FriendsUI/Orders';
-
-
-
+import MainList from './FriendsUI/MainList';
 import ListItemButton from '@mui/material/ListItemButton';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import ListSubheader from '@mui/material/ListSubheader';
-import DashboardIcon from '@mui/icons-material/Dashboard';
-import ShoppingCartIcon from '@mui/icons-material/ShoppingCart';
 import PeopleIcon from '@mui/icons-material/People';
-import BarChartIcon from '@mui/icons-material/BarChart';
-import LayersIcon from '@mui/icons-material/Layers';
-import AssignmentIcon from '@mui/icons-material/Assignment';
 import PeopleOutlineIcon from '@mui/icons-material/PeopleOutline';
 import PersonAddAltIcon from '@mui/icons-material/PersonAddAlt';
-
+import { useState,useEffect } from 'react';
 
 
 function Copyright(props) {
   return (
     <Typography variant="body2" color="text.secondary" align="center" {...props}>
       {'Copyright © '}
-      <Link color="inherit" href="https://mui.com/">
+      <Link color="inherit" href="#">
         Your Website
       </Link>{' '}
       {new Date().getFullYear()}
@@ -84,6 +69,7 @@ const defaultTheme = createTheme();
 
 export default function FriendsPage() {
   const [open, setOpen] = React.useState(true);
+  const [type,setType] = useState('friends');
   const toggleDrawer = () => {
     setOpen(!open);
   };
@@ -98,34 +84,22 @@ export default function FriendsPage() {
         <ListItemIcon>
           <PeopleIcon />
         </ListItemIcon>
-        <ListItemText primary="Friends" />
+        <ListItemText primary="Friends" onClick={()=>{setType('friends')}}/>
       </ListItemButton>
       <ListItemButton style={styles}>
         <ListItemIcon>
           <PersonAddAltIcon />
         </ListItemIcon>
-        <ListItemText primary="Friend Requests" />
+        <ListItemText primary="Friend Requests" onClick={()=>{setType('requests')}}/>
       </ListItemButton>
       <ListItemButton style={styles}>
         <ListItemIcon>
           <PeopleOutlineIcon />
         </ListItemIcon>
-        <ListItemText primary="Followers" />
+        <ListItemText primary="Followers" onClick={()=>{setType('followers')}}/>
       </ListItemButton>
-
     </React.Fragment>
-  
-);
-
-
-
-
-
-
-
-
-
-
+  );
 
 
   return (
@@ -169,7 +143,7 @@ export default function FriendsPage() {
 
               <Grid item xs={12}>
                 <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
-                  <Orders />
+                  <MainList type={type}/>
                 </Paper>
               </Grid>
             </Grid>
